First working version of separate pyopenms build. Fix pyopenms warnings.

+++ src/pyOpenMS/pxds/StringDataArray.pxd
@@ -21,6 +21,6 @@ cdef extern from "<OpenMS/METADATA/DataArrays.h>" namespace "OpenMS::DataArrays"
 
         Size size() nogil except +
         void resize(size_t n) nogil except +
-        String& operator[](int) nogil except + # wrap-ignore
+        String& operator[](size_t) nogil except + # wrap-ignore
         void clear() nogil except +
         void push_back(String) nogil except +
